Accept an optional device parameter
If provided, we attempt to connect to the AxiDraw on this serial port device (if it exists), regardless of product name. This is useful when running within a Docker container, where we seem to not get the product name.

We also update our Dockerfile to allow passing through this command line argument from the host.